refactor: make `LoginForm` its own file
