+++ src/app/(auth)/login/page.tsx
@@ -1,9 +1,8 @@
-import FormItem from '@/shared/container/ClientAntd/FormItem/FormItem';
-import Input from '@/shared/container/ClientAntd/Input/Input';
-import LoginFormContainer from '../_container/LoginFormContainer/LoginFormContainer';
 import { Button, Divider } from 'antd';
 import GoogleIcon from '@/shared/container/Icon/GoogleIcon';
 import Link from 'next/link';
+import LoginForm from '../_container/Form/LoginForm';
+import LoginFormContainer from '../_container/LoginFormContainer';
 
 export default function LoginPage() {
   return (
@@ -16,37 +15,7 @@ export default function LoginPage() {
         <LoginFormContainer
           className="flex flex-col gap-3 w-full"
           layout="vertical">
-          <FormItem
-            rules={[{ required: true, message: 'Please input your email!' }]}
-            className="my-0"
-            required
-            name="email"
-            label="Email">
-            <Input placeholder="Enter your email here!" />
-          </FormItem>
-
-          <FormItem
-            rules={[{ required: true, message: 'Please input your password!' }]}
-            className="my-0"
-            required
-            name="password"
-            label="Password">
-            <Input type="password" placeholder="Enter your password here!" />
-          </FormItem>
-
-          <Link
-            href="/forgot-password"
-            className="text-right text-ny-primary-500 -mt-1 text-caption-1">
-            Forgot Password?
-          </Link>
-
-          <Button
-            size="large"
-            className="mt-2"
-            type="primary"
-            htmlType="submit">
-            Login
-          </Button>
+          <LoginForm />
         </LoginFormContainer>
 
         <Divider plain className="text-caption-2">
